+++ CHANGELOG.md
@@ -2,8 +2,8 @@
 
 All notable changes to this project will be documented in this file.
 
-The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
-and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
+The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/), and this project adheres
+to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased](https://github.com/Smartesting/gravity-data-collector/compare/v2.2.0...main)
 
@@ -13,6 +13,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Changed
 
+- Type `Traits` renamed more precisely to `SessionTraits` and moved to `types.ts` file
+- Type `TraitValue` renamed more precisely to `SessionTraitValue`
+- Method `sendSessionTraits` now returns a response typed `IdentifySessionResponse` with well-typed error 
+
 ### Deprecated
 
 ### Removed
@@ -20,6 +24,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Fixed
 
 - Stop tracking when receiving errors specified in [config.ts](./src/config.ts)
+
+- Prevent error in method `isKeyAllowedByKeyListeners`
 
 ## [2.2.0](https://github.com/Smartesting/gravity-data-collector/compare/v2.1.11...v2.2.0)